@@ -1,19 +1,18 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
-    ext.kotlin_version = '1.2.71'
+    ext.kotlin_version = '1.3.21'
     repositories {
         mavenCentral()
         google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.1'
-        classpath 'com.jakewharton:butterknife-gradle-plugin:8.5.1'
-        classpath 'com.dicedmelon.gradle:jacoco-android:0.1.2'
-        classpath 'org.codehaus.groovy:groovy-all:2.4.12'
+        classpath 'com.android.tools.build:gradle:3.4.0'
+        classpath 'com.dicedmelon.gradle:jacoco-android:0.1.4'
+        classpath 'org.codehaus.groovy:groovy-all:2.4.15'
         classpath 'com.chaitanyapramod.gradle:findbugs-android:1.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'com.google.gms:google-services:4.0.1'
+        classpath 'com.google.gms:google-services:4.2.0'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
