@@ -1,6 +1,7 @@
 package ar.com.wolox.android.cookbook.tests
 
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
+import androidx.test.annotation.UiThreadTest
 import ar.com.wolox.android.cookbook.WolmoTestActivity
 import org.junit.Before
 
@@ -8,6 +9,7 @@ abstract class WolmoFragmentTest<V>(val fragment: V)
     : WolmoActivityTest<WolmoTestActivity>(WolmoTestActivity::class.java) {
 
     @Before
+    @UiThreadTest
     fun setupWolmoFragmentTest() {
         activity.initFragment(fragment as Fragment)
     }
